+++ CHANGELOG.md
@@ -15,6 +15,8 @@ Recommendation: for ease of reading, use the following order:
 ### Changed
 - ObjectRepositoryS3`: can now support external address configuration
 - `kamu repo list`: supports all types of output
+### Fixed
+- `kamu push`: show correct error if server failed to store data
 
 ## [0.206.0] - 2024-10-22
 ### Added